+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+
+- New methods in AIProvider base class:
+  - Add DefaultServerUrl property
+  - Added CallApi method to AIProvider base class supporting GET/POST/DELETE/PATCH
+  - Added RetrieveAvailableModels method to AIProvider base class with default to empty list
+- Implemented RetrieveAvailableModels, CallApi and DefaultServerUrl to existing providers (MistralAIProvider, OpenAIProvider, and DeepSeekProvider).
+
 ### Changed
 
 - Update providersResources access modifiers from public to internal